[REFACTOR/#314] 웹 페이지 로드 방식을 웹뷰에서 외부 브라우저 앱 사용으로 개선했습니다.
